Some minor changes were made

+++ templates/manageusers.html.php
@@ -19,9 +19,9 @@
             <td><?= htmlspecialchars($record->email, ENT_QUOTES, 'UTF-8'); ?></td>
             <td>
               <a href="userprofile.php?id=<?= $record->id ?>" class="btn btn-info">Посмотреть</a>
-              <a href="editprofile.php?id=<?= $record->id ?>" class="btn btn-warning">Редактировать</a>
               
               <?php if ($user->hasPermissions('admin')): ?>
+                <a href="editprofile.php?id=<?= $record->id ?>" class="btn btn-warning">Редактировать</a>
                 <?php if ($record->groupid != 2): ?>
                   <a href="deleteuser.php?id=<?= $record->id ?>" class="btn btn-danger" onclick="return confirm('Вы уверены?');">Удалить</a>
                 <?php endif; ?>
@@ -34,11 +34,14 @@
               <?php endif; ?>
               
               <?php if ($user->hasPermissions('moderator')): ?>
-                <?php // if ($user->hasPermissions('admin') == false): ?>
+                <?php if ($user->hasPermissions('admin') == false): ?>
+                  <?php if ($record->groupid != 2): ?>
+                    <a href="editprofile.php?id=<?= $record->id ?>" class="btn btn-warning">Редактировать</a>
+                  <?php endif; ?>
                   <?php if ($record->groupid == 1): ?>
                     <a href="deleteuser.php?id=<?= $record->id ?>" class="btn btn-danger" onclick="return confirm('Вы уверены?');">Удалить</a>
                   <?php endif; ?>
-                <?php // endif; ?>
+                <?php endif; ?>
               <?php endif; ?>
             </td>
           </tr>
